Add multiple providers
