+++ app/src/main/java/com/rootstrap/android/metrics/base/BaseAnalytics.kt
@@ -2,13 +2,9 @@ package com.rootstrap.android.metrics.base
 
 /**
  * Add custom metrics
- * How use:
- * in the Application class -> onCreate
- * MyMetrics.with(applicationContext)
- * then use in the place you want to track the event (sample):
- * appMetric.logIn()
  * */
-interface AppMetrics {
+interface BaseAnalytics {
+    fun addProvider(provider: Provider)
     fun logOut()
     fun logIn()
     fun signUp()
